core: use only one ID attribute in hsm-wrapper

+++ lib/include/irritator/modeling.hpp
@@ -22,11 +22,11 @@ namespace irt {
 enum class port_id : u32;
 enum class input_connection_id : u32;
 enum class output_connection_id : u32;
-enum class component_id : u64;
-enum class hsm_component_id : u64;
-enum class generic_component_id : u64;
-enum class graph_component_id : u64;
-enum class grid_component_id : u64;
+enum class component_id : u32;
+enum class hsm_component_id : u32;
+enum class generic_component_id : u32;
+enum class graph_component_id : u32;
+enum class grid_component_id : u32;
 enum class tree_node_id : u64;
 enum class description_id : u64;
 enum class child_id : u32;
